BE/requirements.txt
langgraph==0.2.28
langchain==0.3.1
langchain-openai==0.2.1
langchain-community==0.3.1
chromadb==0.5.5
sentence-transformers==3.1.1
spacy==3.7.6
fastapi==0.115.0
uvicorn[standard]==0.31.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
numpy==1.26.4
scikit-learn==1.5.2
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
aiofiles==24.1.0
python-multipart==0.0.12
